fuzz: differential libFuzzer harness + the two bugs it found
`make fuzz` (clang) builds test/fuzz_pipeline.c: parse -> expand ->
lower -> optimize at -O0 and -O -> interpret both under a small step
cap -> abort if any register disagrees. So it hunts optimizer
miscompiles as well as crashes/UB, with examples/ seeding the corpus.
Devices are not wired up, so fuzzed programs can't do I/O.

Found within two minutes: loader expansion is unbounded. Mutually-
calling funcs multiply the AST every round (2^32 nodes before the
depth limit can fire), and a value seed of N emits N inc nodes -- both
exhausted memory. Capped at 131072 expanded statements / seed units
(200x the largest in-tree example) with clean errors, locked in by two
loader tests.

Also fixes int32 overflow (UB) in resolve_jump for relative offsets
near INT32_MAX -- the other thing the fuzzer would have hit instantly.
A 2-minute run now sustains ~370 exec/s, 3.1k edges, zero artifacts.

+++ src/ir.c
@@ -204,13 +204,14 @@ static BcResult resolve_jump(Arena *arena, const Jump *jump, uint32_t current_ad
         break;
 
     case JUMP_OFFSET: {
-        int32_t target = (int32_t)current_addr + jump->offset;
+        // 64-bit arithmetic: an offset near INT32_MAX must clamp, not overflow.
+        int64_t target = (int64_t)current_addr + jump->offset;
         if (target < 0) target = 0;
         // A past-the-end offset means halt. Clamp to the implicit trailing END
         // so every backend agrees: the interpreter treated pc >= inst_count as
         // a silent halt, but the QBE backend emitted jumps to nonexistent
         // blocks and the optimizer had to bounds-check on its own.
-        if ((uint32_t)target >= inst_count) target = (int32_t)inst_count - 1;
+        if (target >= (int64_t)inst_count) target = (int64_t)inst_count - 1;
         return BC_OK((void *)(uintptr_t)target);
     }
     }
